test(endpoint-posts): use node test runner
