+++ packages/endpoint-posts/test/integration/200-posts.js
@@ -1,0 +1,23 @@
+import { strict as assert } from "node:assert";
+import { after, describe, it } from "node:test";
+import supertest from "supertest";
+import { JSDOM } from "jsdom";
+import { testServer } from "@indiekit-test/server";
+import { testCookie } from "@indiekit-test/session";
+
+const server = await testServer();
+const request = supertest.agent(server);
+
+describe("endpoint-posts GET /posts", () => {
+  it("Returns list of published posts", async () => {
+    const response = await request.get("/posts").set("cookie", testCookie());
+    const dom = new JSDOM(response.text);
+    const result = dom.window.document.querySelector("title").textContent;
+
+    assert.equal(result, "Published posts - Test configuration");
+  });
+
+  after(() => {
+    server.close(() => process.exit(0));
+  });
+});
