--- packages/endpoint-posts/tests/integration/200-get-new.js
@@ -1,19 +1,0 @@
-import test from "ava";
-import supertest from "supertest";
-import { JSDOM } from "jsdom";
-import { testServer } from "@indiekit-test/server";
-import { testCookie } from "@indiekit-test/session";
-
-test("Returns create new post page", async (t) => {
-  const server = await testServer();
-  const request = supertest.agent(server);
-  const response = await request
-    .get("/posts/new")
-    .set("cookie", [testCookie()]);
-  const dom = new JSDOM(response.text);
-  const result = dom.window.document.querySelector("title").textContent;
-
-  t.is(result, "What type of post do you want to create? - Test configuration");
-
-  server.close(t);
-});
