--- packages/endpoint-posts/tests/integration/302-post-create.js
@@ -1,24 +1,0 @@
-import test from "ava";
-import supertest from "supertest";
-import { mockAgent } from "@indiekit-test/mock-agent";
-import { testServer } from "@indiekit-test/server";
-import { testCookie } from "@indiekit-test/session";
-
-await mockAgent("endpoint-micropub");
-
-test("Creates post and redirects to posts page", async (t) => {
-  const server = await testServer();
-  const request = supertest.agent(server);
-  const result = await request
-    .post("/posts/create")
-    .type("form")
-    .set("cookie", [testCookie()])
-    .send({ type: "entry" })
-    .send({ content: "Foobar" })
-    .send({ published: "2023-08-28T12:30" });
-
-  t.is(result.status, 302);
-  t.regex(result.text, /Found. Redirecting to \/posts\?success/);
-
-  server.close(t);
-});
